POS has very very very very lite actionManager capabilities
