+++ addons/point_of_sale/static/src/js/main.js
@@ -4,20 +4,10 @@ odoo.define('point_of_sale.main', function (require) {
     const env = require('web.env');
     const Chrome = require('point_of_sale.Chrome');
     const Registries = require('point_of_sale.Registries');
+    const PosRoot = require('point_of_sale.owlPosRoot');
 
     owl.config.mode = env.isDebug() ? 'dev' : 'prod';
     owl.Component.env = env;
-
-    class PosRoot extends owl.Component {
-        mounted() {
-            odoo.isReady = true;
-            this.env.bus.trigger('web-client-mounted');
-            super.mounted();
-        }
-    };
-    PosRoot.template = owl.tags.xml`
-        <Chrome t-att-class="'o_action_manager'"/>
-    `;
 
     const posRoot = new PosRoot(null);
 
@@ -32,9 +22,4 @@ odoo.define('point_of_sale.main', function (require) {
 
     startPosApp();
     return posRoot;
-});
-
-odoo.define('root.widget', function (require) {
-    "use strict";
-    return null;
 });
