+++ addons/point_of_sale/views/pos_assets_common.xml
@@ -22,6 +22,7 @@
     <script type="text/javascript" src="/point_of_sale/static/src/js/keyboard.js"></script>
     <script type="text/javascript" src="/point_of_sale/static/src/js/barcode_reader.js"></script>
     <script type="text/javascript" src="/point_of_sale/static/src/js/Gui.js"></script>
+    <script type="text/javascript" src="/point_of_sale/static/src/js/owl_pos_root.js"></script>
     <script type="text/javascript" src="/point_of_sale/static/src/js/PopupControllerMixin.js"></script>
     <script type="text/javascript" src="/point_of_sale/static/src/js/Chrome.js"></script>
     <script type="text/javascript" src="/point_of_sale/static/src/js/devices.js"></script>
